Add: component select language and theme for related login pages
