+++ frontend/src/pages/login/AuthVerification.tsx
@@ -1,15 +1,13 @@
 import React, { useState } from 'react';
-import { useTheme } from '@/components/ThemeProvider';
 import { useTranslation } from 'react-i18next';
 import { useNavigate } from 'react-router-dom';
 import { Button } from '@/components/ui/button';
 import { Card, CardHeader, CardTitle, CardDescription, CardContent } from '@/components/ui/card';
-import { Select, SelectTrigger, SelectValue, SelectContent, SelectItem } from '@/components/ui/select';
 import { InputOTP, InputOTPGroup, InputOTPSlot } from '@/components/ui/input-otp';
+import { SelectLanguageAndTheme } from "@/components/SelectLanguageAndTheme.tsx";
 
 const AuthVerification = () => {
-	const { themeMode, setThemeMode } = useTheme();
-	const { t, i18n } = useTranslation();
+	const { t } = useTranslation();
 	const navigate = useNavigate();
 
 	const [otpValue, setOtpValue] = useState<string>('');
@@ -20,44 +18,9 @@ const AuthVerification = () => {
 		navigate('/home', { replace: true });
 	};
 
-	const handleLanguageChange = (language: string) => {
-		i18n
-			.changeLanguage(language)
-			.then(() => {
-				localStorage.setItem('language', language);
-			})
-			.catch((error) => {
-				console.error('Failed to change language:', error);
-			});
-	};
-
 	return (
 		<div className="flex h-screen">
-			<div className="absolute top-4 right-4 flex space-x-4">
-				<Select value={themeMode} onValueChange={setThemeMode}>
-					<SelectTrigger className="w-[120px]">
-						<SelectValue placeholder={t('Accessibility.Theme')}/>
-					</SelectTrigger>
-					<SelectContent>
-						<SelectItem value="light">{t('Accessibility.Light')}</SelectItem>
-						<SelectItem value="dark">{t('Accessibility.Dark')}</SelectItem>
-						<SelectItem value="system">{t('Accessibility.System')}</SelectItem>
-					</SelectContent>
-				</Select>
-
-				<Select
-					value={i18n.language}
-					onValueChange={(value) => handleLanguageChange(value)}
-				>
-					<SelectTrigger className="w-[120px]">
-						<SelectValue placeholder={t('Accessibility.Language')}/>
-					</SelectTrigger>
-					<SelectContent>
-						<SelectItem value="en">{t('Accessibility.English')}</SelectItem>
-						<SelectItem value="fr">{t('Accessibility.French')}</SelectItem>
-					</SelectContent>
-				</Select>
-			</div>
+			<SelectLanguageAndTheme/>
 
 			<div className="w-full flex items-center justify-center">
 				<Card className="w-full max-w-md">
