+++ frontend/src/pages/login/Register.tsx
@@ -5,10 +5,9 @@ import { Label } from '@/components/ui/label';
 import { Button } from '@/components/ui/button';
 import { CheckIcon, Cross1Icon, EyeOpenIcon, EyeNoneIcon } from '@radix-ui/react-icons';
 import axios from 'axios';
-import { useTheme } from '@/components/ThemeProvider';
 import { useTranslation } from 'react-i18next';
 import { Card, CardHeader, CardTitle, CardDescription, CardContent } from '@/components/ui/card';
-import { Select, SelectTrigger, SelectValue, SelectContent, SelectItem } from '@/components/ui/select';
+import { SelectLanguageAndTheme } from "@/components/SelectLanguageAndTheme.tsx";
 
 const PasswordCriteria = ({ label, isValid }: { label: string; isValid: boolean }) => (
   <div className="flex items-center">
@@ -22,8 +21,7 @@ const PasswordCriteria = ({ label, isValid }: { label: string; isValid: boolean 
 );
 
 const Register = () => {
-  const { themeMode, setThemeMode } = useTheme();
-  const { t, i18n } = useTranslation();
+  const { t } = useTranslation();
 
   const [formData, setFormData] = useState({
     username: '',
@@ -125,44 +123,9 @@ const Register = () => {
     }
   };
 
-  const handleLanguageChange = (language: string) => {
-    i18n
-      .changeLanguage(language)
-      .then(() => {
-        localStorage.setItem('language', language);
-      })
-      .catch((error) => {
-        console.error('Failed to change language:', error);
-      });
-  };
-
   return (
     <div className="flex h-screen">
-      <div className="absolute top-4 right-4 flex space-x-4">
-        <Select value={themeMode} onValueChange={setThemeMode}>
-          <SelectTrigger className="w-[120px]">
-            <SelectValue placeholder={t('Accessibility.Theme')}/>
-          </SelectTrigger>
-          <SelectContent>
-            <SelectItem value="light">{t('Accessibility.Light')}</SelectItem>
-            <SelectItem value="dark">{t('Accessibility.Dark')}</SelectItem>
-            <SelectItem value="system">{t('Accessibility.System')}</SelectItem>
-          </SelectContent>
-        </Select>
-
-        <Select
-          value={i18n.language}
-          onValueChange={(value) => handleLanguageChange(value)}
-        >
-          <SelectTrigger className="w-[120px]">
-            <SelectValue placeholder={t('Accessibility.Language')}/>
-          </SelectTrigger>
-          <SelectContent>
-            <SelectItem value="en">{t('Accessibility.English')}</SelectItem>
-            <SelectItem value="fr">{t('Accessibility.French')}</SelectItem>
-          </SelectContent>
-        </Select>
-      </div>
+      <SelectLanguageAndTheme/>
 
       <div className="w-full flex items-center justify-center">
         <Card className="w-full max-w-md">
